fix: Refactor images

+++ src/seed/seed.service.ts
@@ -1,4 +1,6 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
+
 import { AuthService } from 'src/auth/auth.service';
 import { ProductsService } from 'src/products/products.service';
 import { initialProducts, initialUsers } from './data/seed-data';
@@ -7,12 +9,17 @@ import { User } from 'src/auth/entities/user.entity';
 @Injectable()
 export class SeedService {
   constructor (
+    private readonly configService: ConfigService,
     private readonly productsService: ProductsService,
-
-    private readonly authService: AuthService    
+    private readonly authService: AuthService
   ) { }
 
   async executeSeed() {
+    const environment = this.configService.get<string>('NODE_ENV') ?? 'development';
+    if (environment !== 'development') {
+      throw new ForbiddenException('Endpoint ejecutable solo en ambiente de pruebas');
+    }
+
     await this.deleteTables();
 
     await Promise.all(
